іbackward condition

+++ app/main.py
@@ -10,10 +10,6 @@ def move_file(command: str) -> None:
             f" But you have typed {input_command} instead!"
         )
 
-    if ("/" or "\\") not in path_file_out:
-        os.rename(path_file_in, path_file_out)
-        return
-
     output_directory_file_path = os.path.abspath(path_file_out)
     output_directory_path = os.path.split(output_directory_file_path)[0]
     try:
@@ -24,9 +20,12 @@ def move_file(command: str) -> None:
     else:
         os.makedirs(output_directory_path)
 
-    with (
-        open(path_file_in) as file_in,
-        open(output_directory_file_path, "w") as file_out,
-    ):
-        file_out.write(file_in.read())
-    os.remove(path_file_in)
+    if ("/" or "\\") in path_file_out:
+        with (
+            open(path_file_in) as file_in,
+            open(output_directory_file_path, "w") as file_out,
+        ):
+            file_out.write(file_in.read())
+        os.remove(path_file_in)
+    else:
+        os.rename(path_file_in, path_file_out)
